fix(build): rename package to `halper`
