+++ src/halper/utils/__init__.py
@@ -6,9 +6,9 @@ from .logging import InterceptHandler, instantiate_logger  # isort:skip
 from .helpers import get_tldr_command, list_commands
 
 __all__ = [
+    "InterceptHandler",
     "console",
     "get_tldr_command",
     "instantiate_logger",
-    "InterceptHandler",
     "list_commands",
 ]
